Removed Games.js, fixed min-max player issue
