--- src/components/GamesPageContainer.js
@@ -1,33 +1,0 @@
-import React, { useContext } from 'react';
-import '../css/GamesPageContainer.css';
-import GameInfoCard from './GameInfoCard'
-import { GameContext } from '../context/GameContext';
-
-const GamesPageContainer = () => {
-	const { games } = useContext(GameContext)
-
-	console.log(games)
-
-	const gameCards = games.map(game => {
-    return (
-      <GameInfoCard
-        name={game.name}
-        image={game.img}
-        id={game.id}
-        key={game.id}
-				price={game.price}
-				minPlayers={game.min_players}
-				maxPlayers={game.max_players}
-				url={game.url}
-      />
-    )
-  })
-
-	return (
-		<section className='game-page-container'>
-			{gameCards}
-		</section>
-	)
-}
-
-export default GamesPageContainer;
