+++ src/css/Games.css
@@ -1,0 +1,6 @@
+.game-page-container {
+  padding: 35px;
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(250px, auto));
+  gap: 1% 2%;
+  }
