--- src/css/GamesPageContainer.css
@@ -1,6 +1,0 @@
-.game-page-container {
-  padding: 35px;
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(250px, auto));
-  gap: 1% 2%;
-  }
